@@ -5,7 +5,7 @@ describe Vmstat::NetworkInterface do
     let(:network) { Vmstat.network_interfaces }
 
     it "should return the enthernet and loopback network data as an array" do
-      network.should be_a(Array)
+      expect(network).to be_a(Array)
     end
 
     context "loopback device" do
@@ -13,25 +13,48 @@ describe Vmstat::NetworkInterface do
       subject { loopback }
 
       it "should be a vmstat network interface object" do
-        should be_a(described_class)
+        is_expected.to be_a(described_class)
       end
 
       context "methods" do
-        it { should respond_to(:in_bytes) }
-        it { should respond_to(:out_bytes) }
-        it { should respond_to(:in_errors) }
-        it { should respond_to(:out_errors) }
-        it { should respond_to(:in_drops) }
-        it { should respond_to(:type) }
+        it { is_expected.to respond_to(:in_bytes) }
+        it { is_expected.to respond_to(:out_bytes) }
+        it { is_expected.to respond_to(:in_errors) }
+        it { is_expected.to respond_to(:out_errors) }
+        it { is_expected.to respond_to(:in_drops) }
+        it { is_expected.to respond_to(:type) }
       end
 
       context "content" do
-        its(:in_bytes) { should be_a_kind_of(Numeric) }
-        its(:out_bytes) { should be_a_kind_of(Numeric) }
-        its(:in_errors) { should be_a_kind_of(Numeric) }
-        its(:out_errors) { should be_a_kind_of(Numeric) }
-        its(:in_drops) { should be_a_kind_of(Numeric) }
-        its(:type) { should be_a_kind_of(Numeric) }
+        describe '#in_bytes' do
+          subject { super().in_bytes }
+          it { is_expected.to be_a_kind_of(Numeric) }
+        end
+
+        describe '#out_bytes' do
+          subject { super().out_bytes }
+          it { is_expected.to be_a_kind_of(Numeric) }
+        end
+
+        describe '#in_errors' do
+          subject { super().in_errors }
+          it { is_expected.to be_a_kind_of(Numeric) }
+        end
+
+        describe '#out_errors' do
+          subject { super().out_errors }
+          it { is_expected.to be_a_kind_of(Numeric) }
+        end
+
+        describe '#in_drops' do
+          subject { super().in_drops }
+          it { is_expected.to be_a_kind_of(Numeric) }
+        end
+
+        describe '#type' do
+          subject { super().type }
+          it { is_expected.to be_a_kind_of(Numeric) }
+        end
       end
     end
   end
