@@ -5,23 +5,23 @@ describe Vmstat do
     let(:boot_time) { Vmstat.boot_time }
 
     it "should be an array" do
-      boot_time.should be_a(Time)
+      expect(boot_time).to be_a(Time)
     end
 
     it "has to be a time before now" do
-      boot_time.should < Time.now
+      expect(boot_time).to be < Time.now
     end
   end
   
   context "Vmstat#filter_devices" do
     unless travis? # no external ethernet devices on travis ci
       it "should filter ethernet devices" do
-        Vmstat.ethernet_devices.size.should >= 1
+        expect(Vmstat.ethernet_devices.size).to be >= 1
       end
     end
 
     it "should filter loopback devices" do
-      Vmstat.loopback_devices.size.should >= 1
+      expect(Vmstat.loopback_devices.size).to be >= 1
     end
   end
 
@@ -34,7 +34,7 @@ describe Vmstat do
         mem_before = Vmstat.task.resident_size
         10000.times { Vmstat.send(method_name, *args) }
         mem_after = Vmstat.task.resident_size
-        mem_after.should < (mem_before * (1 + percent / 100.0))
+        expect(mem_after).to be < (mem_before * (1 + percent / 100.0))
       end
     end
 
