@@ -6,21 +6,36 @@ describe Vmstat::Task do
     subject { task }
 
     it "should be a vmstat task object" do
-      should be_a(described_class)
+      is_expected.to be_a(described_class)
     end
 
     context "methods" do
-      it { should respond_to(:virtual_size) }
-      it { should respond_to(:resident_size) }
-      it { should respond_to(:user_time_ms) }
-      it { should respond_to(:system_time_ms) }
+      it { is_expected.to respond_to(:virtual_size) }
+      it { is_expected.to respond_to(:resident_size) }
+      it { is_expected.to respond_to(:user_time_ms) }
+      it { is_expected.to respond_to(:system_time_ms) }
     end
     
     context "content" do
-      its(:virtual_size) { should be_a_kind_of(Numeric) }
-      its(:resident_size) { should be_a_kind_of(Numeric) }
-      its(:user_time_ms) { should be_a_kind_of(Numeric) }
-      its(:system_time_ms) { should be_a_kind_of(Numeric) }
+      describe '#virtual_size' do
+        subject { super().virtual_size }
+        it { is_expected.to be_a_kind_of(Numeric) }
+      end
+
+      describe '#resident_size' do
+        subject { super().resident_size }
+        it { is_expected.to be_a_kind_of(Numeric) }
+      end
+
+      describe '#user_time_ms' do
+        subject { super().user_time_ms }
+        it { is_expected.to be_a_kind_of(Numeric) }
+      end
+
+      describe '#system_time_ms' do
+        subject { super().system_time_ms }
+        it { is_expected.to be_a_kind_of(Numeric) }
+      end
     end
   end
 end if RUBY_PLATFORM =~ /darwin|linux/
